chore(types): consolidate and barrel exports
