+++ jsconfig.json
@@ -9,7 +9,8 @@
       "~utils/*": ["utils/*"],
       "~data/*": ["data/*"],
       "~config/*": ["config/*"],
-      "~types/*": ["src/types/*"]
+      "~types": ["types"],
+      "~types/*": ["types/*"]
     },
     "jsx": "preserve",
     "jsxImportSource": "astro"
