+++ src/components/StatsMilestonePage.astro
@@ -3,7 +3,7 @@ import DescriptionText from '~components/DescriptionText.astro';
 import Heading from '~components/Heading.astro';
 import WordLink from '~components/WordLink.astro';
 import Layout from '~layouts/Layout.astro';
-import type { WordMilestoneItem } from '~types/word';
+import type { WordMilestoneItem } from '~types';
 import { getPageMetadata } from '~astro-utils/page-metadata.ts';
 import { STRUCTURED_DATA_TYPE } from '~astro-utils/schema-utils.ts';
 import { formatWordCount } from '~utils/text-utils';
