+++ src/components/StatsWordListPage.astro
@@ -3,7 +3,7 @@ import DescriptionText from '~components/DescriptionText.astro';
 import Heading from '~components/Heading.astro';
 import WordSection from '~components/WordSection.astro';
 import Layout from '~layouts/Layout.astro';
-import type { WordData } from '~types/word';
+import type { WordData } from '~types';
 import { getPageMetadata } from '~astro-utils/page-metadata.ts';
 import { STRUCTURED_DATA_TYPE } from '~astro-utils/schema-utils.ts';
 import { formatWordCount } from '~utils/text-utils';
