+++ src/components/StructuredData.astro
@@ -1,6 +1,6 @@
 ---
 
-import type { CollectionPageSchema, DefinedTermSchema, WebSiteSchema, WordSchemaData } from '~types/schema';
+import type { CollectionPageSchema, DefinedTermSchema, WebSiteSchema, WordSchemaData } from '~types';
 import { STRUCTURED_DATA_TYPE, type StructuredDataType } from '~astro-utils/schema-utils.ts';
 import { seoConfig } from '~astro-utils/seo-utils.ts';
 import { getFullUrl } from '~astro-utils/url-utils';
