+++ src/components/Word.astro
@@ -1,7 +1,7 @@
 ---
 import Heading from '~components/Heading.astro';
 import WordDescription from '~components/WordDescription.astro';
-import type { WordData } from '~types/word';
+import type { WordData } from '~types';
 import { formatDate } from '~utils/date-utils';
 import { getWordDetails } from '~astro-utils/word-data-utils';
