+++ src/components/WordDescription.astro
@@ -1,5 +1,5 @@
 ---
-import type { SourceMeta } from '~types/common';
+import type { SourceMeta } from '~types';
 
 interface Props {
   partOfSpeech: string;
