+++ src/components/WordPage.astro
@@ -2,7 +2,7 @@
 import WordComponent from '~components/Word.astro';
 import WordNav from '~components/WordNav.astro';
 import Layout from '~layouts/Layout.astro';
-import type { WordData } from '~types/word';
+import type { WordData } from '~types';
 import { STRUCTURED_DATA_TYPE } from '~astro-utils/schema-utils.ts';
 import { getMetaDescription } from '~astro-utils/seo-utils.ts';
 import { extractWordDefinition } from '~astro-utils/word-data-utils';
