+++ src/components/WordSection.astro
@@ -1,7 +1,7 @@
 ---
 import SectionHeading from '~components/SectionHeading.astro';
 import WordList from '~components/WordList.astro';
-import type { WordData } from '~types/word';
+import type { WordData } from '~types';
 
 interface Props {
   title?: string;
