+++ src/layouts/Layout.astro
@@ -5,7 +5,7 @@ import '~styles/theme.css';
 import Footer from '~components/Footer.astro';
 import Header from '~components/Header.astro';
 import StructuredData from '~components/StructuredData.astro';
-import type { WordData } from '~types/word';
+import type { WordData } from '~types';
 import { getBuildData } from '~astro-utils/build-utils.ts';
 import { getSocialImageUrl } from '~astro-utils/image-utils';
 import { getMetaDescription, seoConfig } from '~astro-utils/seo-utils.ts';
