+++ src/pages/stats/[stat].astro
@@ -5,7 +5,7 @@ import { generateStatsStaticPaths } from '~astro-utils/static-paths-utils';
 
 export const getStaticPaths = generateStatsStaticPaths;
 
-import type { WordData, WordMilestoneItem } from '~types/word';
+import type { WordData, WordMilestoneItem } from '~types';
 
 const { words, description, template } = Astro.props;
