+++ src/utils/build-utils.ts
@@ -1,4 +1,4 @@
-import type { WordData } from '~types/word';
+import type { WordData } from '~types';
 import { generateWordDataHash } from '~astro-utils/word-data-utils';
 
 export interface BuildData {
