+++ src/utils/seo-utils.ts
@@ -3,7 +3,7 @@
  * Centralized SEO config following Astro best practices
  */
 
-import type { SeoConfig, SeoMetadata, SeoMetadataOptions, SeoMetaDescriptionOptions } from '~types/seo';
+import type { SeoConfig, SeoMetadata, SeoMetadataOptions, SeoMetaDescriptionOptions } from '~types';
 import { getFullUrl } from '~astro-utils/url-utils';
 
 // SEO configuration using environment variables - no fallbacks for security
